+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Chaos rune recipes not producing valid runes
+- Empty pedestals absorbing chaos
 
 ## [3.3.5] - 2019-08-12
 ### Added